requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-location-field==2.7.3
django-ninja==1.3.0
dmgbuild==1.6.4
ds-store==1.3.1
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
load-dotenv==0.1.0
mac-alias==2.2.2
macholib==1.16.3
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uuid==1.30
whitenoise==6.9.0
zope.interface==7.2
